Fix -Wc++11-narrowing in V8.
Bug: chromium:1216696
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2950262
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#75062}

+++ test/cctest/wasm/test-compilation-cache.cc
@@ -74,7 +74,7 @@ class StreamTester {
 };
 
 // Create a valid module such that the bytes depend on {n}.
-ZoneBuffer GetValidModuleBytes(Zone* zone, int n) {
+ZoneBuffer GetValidModuleBytes(Zone* zone, uint8_t n) {
   ZoneBuffer buffer(zone);
   TestSignatures sigs;
   WasmModuleBuilder builder(zone);
